Fix test file imports after database refactor
- I updated import statements and class instantiations in `tests/test_kmer_database.py` and `tests/test_strain_kmer_db.py` to use the new `StrainKmerDatabase` class from `strainr.database`.
- I added a local type alias for `Kmer` in `tests/test_strain_kmer_db.py` as it was previously imported from the now-deleted `kmer_database.py`.
- This commit is a follow-up to the main database code refactoring.
