+++ src/strainr/classify.py
@@ -36,7 +36,7 @@ except ImportError as e:
 # Assuming these local modules are structured correctly within the 'strainr' package
 from strainr.analyze import ClassificationAnalyzer
 from strainr.genomic_types import CountVector, ReadId, StrainIndex
-from strainr.kmer_database import StrainKmerDb
+from strainr.database import StrainKmerDatabase # Updated import
 
 # Type aliases for better readability
 ReadHitResults = List[Tuple[ReadId, CountVector]]
@@ -440,15 +440,15 @@ class KmerClassificationWorkflow:
     def __init__(self, args: CliArgs) -> None:
         """Initialize the workflow with validated arguments."""
         self.args = args
-        self.database: Optional[StrainKmerDb] = None
+        self.database: Optional[StrainKmerDatabase] = None
         self.logger = logging.getLogger(__name__)
         self.logger.info("Initialized KmerClassificationWorkflow")
 
     def _initialize_database(self) -> None:
-        """Loads and initializes the StrainKmerDb."""
+        """Loads and initializes the StrainKmerDatabase."""
         self.logger.info(f"Loading k-mer database from: {self.args.db_path}")
         try:
-            self.database = StrainKmerDb(self.args.db_path)
+            self.database = StrainKmerDatabase(self.args.db_path)
             self.logger.info(
                 f"Database loaded: {self.database.num_kmers} k-mers, "
                 f"{self.database.num_strains} strains, "
